examples/demo06/sample/requirements.txt
asyncio_redis
jinja2
feedparser
sanic
aiocache
pytest-sanic
locustio
gunicorn
